requirements.txt
annotated-types==0.7.0
anyio==4.3.0
attrs==23.2.0
certifi==2024.2.2
colorama==0.4.6
distro==1.9.0
Faker==25.2.0
flake8==7.0.0
flake8-annotations==3.0.1
flake8-quotes==3.4.0
flake8-variables-names==0.0.6
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
mccabe==0.7.0
openai==1.30.3
packaging==24.0
pep8-naming==0.13.3
pluggy==1.5.0
pycodestyle==2.11.1
pydantic==2.7.1
pydantic_core==2.18.2
pyflakes==3.2.0
pytest==8.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.12.0
